fix: fail closed without prompt provenance
