+++ pkg/services/factory_definitions/internal/services/snapshots_portability/capture/capture.go
@@ -99,7 +99,9 @@ func CaptureLoaded(
 	if err != nil {
 		return nil, fmt.Errorf("encode factory snapshot: %w", err)
 	}
-	if spanSource, ok := source.(factorydefinitions.InvocationSensitiveJSONSpanSource); ok {
+	if spanSource, ok := source.(interface {
+		InvocationSensitiveJSONSpans() []factorydefinitions.InvocationSensitiveJSONSpan
+	}); ok {
 		object, err = redactInvocationSensitiveSpans(object, spanSource.InvocationSensitiveJSONSpans())
 		if err != nil {
 			return nil, fmt.Errorf("redact invocation-sensitive Factory snapshot: %w", err)
